fixed thresholding in wrapper and included feature and outcome mappings

+++ periomod/app/_inputprocessor.py
@@ -287,6 +287,7 @@ class InputProcessor:
         else:
             return prediction_output
 
-        prediction_output["prediction"] = prediction_output["prediction"].astype(int)
-        prediction_output["prediction"] = prediction_output["prediction"].map(mapping)
+        prediction_output["prediction"] = (
+            prediction_output["prediction"].astype(int).map(mapping)
+        )
         return prediction_output
